// src/modules/subscription/enums/plan-feature.enum.ts
export enum PlanFeature {
  REVIEWS = 'reviews',
  MARKETPLACE = 'marketplace',
  REGIONAL_HIGHLIGHT = 'regionalHighlight',
  ELITE_BADGE = 'eliteBadge',
  WHATSAPP_NOTIFICATION = 'whatsappNotification',
  REPORT_EXPORT = 'reportExport',
  REPORTS_BASIC = 'reports_basic',
  REPORTS_INTERMEDIATE = 'reports_intermediate',
  REPORTS_ADVANCED = 'reports_advanced',
  CUSTOMIZATION_BASIC = 'customization_basic',
  CUSTOMIZATION_INTERMEDIATE = 'customization_intermediate',
  CUSTOMIZATION_FULL = 'customization_full',
}